Add Justfile, rename bindings/js to bindings/ts
