+++ bin/just
@@ -1,0 +1,1 @@
+.just-1.23.0.pkg
