--- bindings/js/pkg/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "web5-wasm",
-  "version": "0.1.0",
-  "license": "SEE LICENSE IN ../../LICENSE",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/TBD54566975/web5-rs.git"
-  },
-  "files": [
-    "web5_wasm_bg.wasm",
-    "web5_wasm.js",
-    "web5_wasm.d.ts"
-  ],
-  "main": "web5_wasm.js",
-  "homepage": "https://github.com/TBD54566975/web5-rs",
-  "types": "web5_wasm.d.ts"
-}
