--- bindings/js/src/main.ts
@@ -1,1 +1,0 @@
-export * from '../pkg'
