--- bindings/swift/test.sh
@@ -1,5 +1,0 @@
-#/bin/bash
-
-set -e
-
-swift test
