+++ bindings/ts/jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  preset          : 'ts-jest',
+  testEnvironment : 'node',
+}
