+++ bindings/ts/src/main.ts
@@ -1,0 +1,1 @@
+export * from '../pkg'
