--- bindings/uniffi/bind.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-
-set -e
-
-cargo build --release
-
-./kotlin.sh
-./swift.sh
